Update generated files with build 50667 [skip ci]

+++ src/Beta/Microsoft/Graph/Model/ItemActionSet.php
@@ -32,14 +32,15 @@ class ItemActionSet extends Entity
     */
     public function getComment()
     {
-        if (array_key_exists("comment", $this->_propDict)) {
-            if (is_a($this->_propDict["comment"], "\Beta\Microsoft\Graph\Model\CommentAction") || is_null($this->_propDict["comment"])) {
+        if (array_key_exists("comment", $this->_propDict) && !is_null($this->_propDict["comment"])) {
+     
+            if (is_a($this->_propDict["comment"], "\Beta\Microsoft\Graph\Model\CommentAction")) {
                 return $this->_propDict["comment"];
             } else {
                 $this->_propDict["comment"] = new CommentAction($this->_propDict["comment"]);
                 return $this->_propDict["comment"];
-            }
-        }
+            } 
+             }
         return null;
     }
 
@@ -65,14 +66,15 @@ class ItemActionSet extends Entity
     */
     public function getCreate()
     {
-        if (array_key_exists("create", $this->_propDict)) {
-            if (is_a($this->_propDict["create"], "\Beta\Microsoft\Graph\Model\CreateAction") || is_null($this->_propDict["create"])) {
+        if (array_key_exists("create", $this->_propDict) && !is_null($this->_propDict["create"])) {
+     
+            if (is_a($this->_propDict["create"], "\Beta\Microsoft\Graph\Model\CreateAction")) {
                 return $this->_propDict["create"];
             } else {
                 $this->_propDict["create"] = new CreateAction($this->_propDict["create"]);
                 return $this->_propDict["create"];
-            }
-        }
+            } 
+             }
         return null;
     }
 
@@ -98,14 +100,15 @@ class ItemActionSet extends Entity
     */
     public function getDelete()
     {
-        if (array_key_exists("delete", $this->_propDict)) {
-            if (is_a($this->_propDict["delete"], "\Beta\Microsoft\Graph\Model\DeleteAction") || is_null($this->_propDict["delete"])) {
+        if (array_key_exists("delete", $this->_propDict) && !is_null($this->_propDict["delete"])) {
+     
+            if (is_a($this->_propDict["delete"], "\Beta\Microsoft\Graph\Model\DeleteAction")) {
                 return $this->_propDict["delete"];
             } else {
                 $this->_propDict["delete"] = new DeleteAction($this->_propDict["delete"]);
                 return $this->_propDict["delete"];
-            }
-        }
+            } 
+             }
         return null;
     }
 
@@ -131,14 +134,15 @@ class ItemActionSet extends Entity
     */
     public function getEdit()
     {
-        if (array_key_exists("edit", $this->_propDict)) {
-            if (is_a($this->_propDict["edit"], "\Beta\Microsoft\Graph\Model\EditAction") || is_null($this->_propDict["edit"])) {
+        if (array_key_exists("edit", $this->_propDict) && !is_null($this->_propDict["edit"])) {
+     
+            if (is_a($this->_propDict["edit"], "\Beta\Microsoft\Graph\Model\EditAction")) {
                 return $this->_propDict["edit"];
             } else {
                 $this->_propDict["edit"] = new EditAction($this->_propDict["edit"]);
                 return $this->_propDict["edit"];
-            }
-        }
+            } 
+             }
         return null;
     }
 
@@ -164,14 +168,15 @@ class ItemActionSet extends Entity
     */
     public function getMention()
     {
-        if (array_key_exists("mention", $this->_propDict)) {
-            if (is_a($this->_propDict["mention"], "\Beta\Microsoft\Graph\Model\MentionAction") || is_null($this->_propDict["mention"])) {
+        if (array_key_exists("mention", $this->_propDict) && !is_null($this->_propDict["mention"])) {
+     
+            if (is_a($this->_propDict["mention"], "\Beta\Microsoft\Graph\Model\MentionAction")) {
                 return $this->_propDict["mention"];
             } else {
                 $this->_propDict["mention"] = new MentionAction($this->_propDict["mention"]);
                 return $this->_propDict["mention"];
-            }
-        }
+            } 
+             }
         return null;
     }
 
@@ -197,14 +202,15 @@ class ItemActionSet extends Entity
     */
     public function getMove()
     {
-        if (array_key_exists("move", $this->_propDict)) {
-            if (is_a($this->_propDict["move"], "\Beta\Microsoft\Graph\Model\MoveAction") || is_null($this->_propDict["move"])) {
+        if (array_key_exists("move", $this->_propDict) && !is_null($this->_propDict["move"])) {
+     
+            if (is_a($this->_propDict["move"], "\Beta\Microsoft\Graph\Model\MoveAction")) {
                 return $this->_propDict["move"];
             } else {
                 $this->_propDict["move"] = new MoveAction($this->_propDict["move"]);
                 return $this->_propDict["move"];
-            }
-        }
+            } 
+             }
         return null;
     }
 
@@ -230,14 +236,15 @@ class ItemActionSet extends Entity
     */
     public function getRename()
     {
-        if (array_key_exists("rename", $this->_propDict)) {
-            if (is_a($this->_propDict["rename"], "\Beta\Microsoft\Graph\Model\RenameAction") || is_null($this->_propDict["rename"])) {
+        if (array_key_exists("rename", $this->_propDict) && !is_null($this->_propDict["rename"])) {
+     
+            if (is_a($this->_propDict["rename"], "\Beta\Microsoft\Graph\Model\RenameAction")) {
                 return $this->_propDict["rename"];
             } else {
                 $this->_propDict["rename"] = new RenameAction($this->_propDict["rename"]);
                 return $this->_propDict["rename"];
-            }
-        }
+            } 
+             }
         return null;
     }
 
@@ -263,14 +270,15 @@ class ItemActionSet extends Entity
     */
     public function getRestore()
     {
-        if (array_key_exists("restore", $this->_propDict)) {
-            if (is_a($this->_propDict["restore"], "\Beta\Microsoft\Graph\Model\RestoreAction") || is_null($this->_propDict["restore"])) {
+        if (array_key_exists("restore", $this->_propDict) && !is_null($this->_propDict["restore"])) {
+     
+            if (is_a($this->_propDict["restore"], "\Beta\Microsoft\Graph\Model\RestoreAction")) {
                 return $this->_propDict["restore"];
             } else {
                 $this->_propDict["restore"] = new RestoreAction($this->_propDict["restore"]);
                 return $this->_propDict["restore"];
-            }
-        }
+            } 
+             }
         return null;
     }
 
@@ -296,14 +304,15 @@ class ItemActionSet extends Entity
     */
     public function getShare()
     {
-        if (array_key_exists("share", $this->_propDict)) {
-            if (is_a($this->_propDict["share"], "\Beta\Microsoft\Graph\Model\ShareAction") || is_null($this->_propDict["share"])) {
+        if (array_key_exists("share", $this->_propDict) && !is_null($this->_propDict["share"])) {
+     
+            if (is_a($this->_propDict["share"], "\Beta\Microsoft\Graph\Model\ShareAction")) {
                 return $this->_propDict["share"];
             } else {
                 $this->_propDict["share"] = new ShareAction($this->_propDict["share"]);
                 return $this->_propDict["share"];
-            }
-        }
+            } 
+             }
         return null;
     }
 
@@ -329,14 +338,15 @@ class ItemActionSet extends Entity
     */
     public function getVersion()
     {
-        if (array_key_exists("version", $this->_propDict)) {
-            if (is_a($this->_propDict["version"], "\Beta\Microsoft\Graph\Model\VersionAction") || is_null($this->_propDict["version"])) {
+        if (array_key_exists("version", $this->_propDict) && !is_null($this->_propDict["version"])) {
+     
+            if (is_a($this->_propDict["version"], "\Beta\Microsoft\Graph\Model\VersionAction")) {
                 return $this->_propDict["version"];
             } else {
                 $this->_propDict["version"] = new VersionAction($this->_propDict["version"]);
                 return $this->_propDict["version"];
-            }
-        }
+            } 
+             }
         return null;
     }
